Move CompanyOverview code into the modules directory
