+++ src/client/modules/Companies/CompanyOverview/TableCards/reducer.js
@@ -1,7 +1,7 @@
 import { camelCase } from 'lodash'
 
-import { OVERVIEW__COMPANY_INVESTMENT_WON_COUNT } from '../../../../../client/actions'
-import { STAGE_ACTIVE } from '../../../../../client/modules/Investments/Projects/constants'
+import { OVERVIEW__COMPANY_INVESTMENT_WON_COUNT } from '../../../../actions'
+import { STAGE_ACTIVE } from '../../../Investments/Projects/constants'
 
 const initialState = {
   results: [],
